+++ examcom/src/main/java/com/example/action/model/base/Tb1KoSDeleteAction.java
@@ -56,10 +56,12 @@ public class Tb1KoSDeleteAction extends BaseAction {
             Tb1Ko e = FormValidator.toBean(Tb1Ko.class.getName(), gridRow);
 
             java.util.List<com.example.entity.Tb1Shison> tb1Shisons = e.referTb1Shisons();
-            for (com.example.entity.Tb1Shison tb1Shison : tb1Shisons) {
+            if (tb1Shisons != null) {
+                for (com.example.entity.Tb1Shison tb1Shison : tb1Shisons) {
 
-                if (tb1Shison.delete() != 1) {
-                    throw new OptLockError("error.cant.delete");
+                    if (tb1Shison.delete() != 1) {
+                        throw new OptLockError("error.cant.delete");
+                    }
                 }
             }
